move carousel to seperate file
